+++ code/nel/tools/3d/ig_lighter/ig_lighter.cpp
@@ -80,7 +80,7 @@ public:
 		pgId= min(pgId, (uint)(BAR_LENGTH-1));
 		sprintf (msg, "\r%s: %s", message, progressbar[pgId]);
 		uint i;
-		for (i=strlen(msg); i<79; i++)
+		for (i=(uint)strlen(msg); i<79; i++)
 			msg[i]=' ';
 		msg[i]=0;
 		printf (msg);
